Widen accepted structures
Before, what was accepted was more constrained:

- no duplicates
- all seeds had to be explicitly given

Both of these have been relaxed.

It is now possible to define the same seed twice. These
are de-duplicated using reference equality. It is still
not acceptable to have two different seeds with the
same ID.

It is also now possible to omit seeds if they are
dependencies of other seeds. E.g. if A>B>C, you can
supply just A. B and C will be found and run.